+++ packages/vendure-plugin-accept-blue/CHANGELOG.md
@@ -1,3 +1,16 @@
+# 1.1.0 (2024-03-12)
+
+- Refactored order line custom field `subscriptionIds` to `acceptBlueSubscriptionIds` to avoid conflicts with for example the Stripe Plugin
+
+# 1.0.1 (2024-03-08)
+
+- Correctly check given input types
+
+# 1.0.0 (2024-03-07)
+
+- Always save payment methods and create recurring schedules and one time charges with a saved payment method
+- Allow payment for order with a saved payment method ID
+
 # 0.1.0 (2024-02-20)
 
 - Implement Charge Transaction with e2e tests